add front, dragon, test, fix some back & api issue

+++ src/main/java/com/rolandhuon/clashofclans/dto/PlayerDto.java
@@ -2,7 +2,7 @@ package com.rolandhuon.clashofclans.dto;
 
 import com.rolandhuon.clashofclans.model.Player;
 
-public record PlayerDto(Long id, String name, int level, int gold, int elixir, int darkElixir, int trophies) {
+public record PlayerDto(Long id, String name, int level, long gold, long elixir, long darkElixir, int trophies) {
 
     public static PlayerDto from(Player player){
         return new PlayerDto(player.getId(), player.getName(), player.getLevel(),
