+++ src/main/java/com/rolandhuon/clashofclans/dto/RaidResultDto.java
@@ -4,7 +4,7 @@ import com.rolandhuon.clashofclans.service.RaidOutcome;
 
 public record RaidResultDto(int destructionPercentage, int stars, int turns,
                             int survivingTroops, boolean villageDestroyed,
-                            int lootedGold, int lootedElixir, int lootedDarkElixir,
+                            long lootedGold, long lootedElixir, long lootedDarkElixir,
                             int trophyChange, int attackerTrophies, int defenderTrophies) {
 
     public static RaidResultDto from(RaidOutcome outcome) {
