+++ src/main/java/com/rolandhuon/clashofclans/service/InsufficientResourcesException.java
@@ -2,7 +2,7 @@ package com.rolandhuon.clashofclans.service;
 
 public class InsufficientResourcesException extends RuntimeException {
 
-    public InsufficientResourcesException(String resource, int available, int required) {
+    public InsufficientResourcesException(String resource, long available, long required) {
         super("Not enough " + resource + ": " + available + " available, " + required + " required.");
     }
 }
